+++ packages/rest-api/src/constants/index.ts
@@ -6,6 +6,7 @@ export const VALID_BRIDGE_MODULES = [
   'SynapseBridge',
   'SynapseCCTP',
   'SynapseRFQ',
+  'Gas.zip',
 ]
 
 export const ZeroAddress = '0x0000000000000000000000000000000000000000'